requirements.txt
annotated-types==0.7.0
azure-ai-documentintelligence==1.0.2
azure-core==1.38.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
idna==3.11
isodate==0.7.2
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
python-dotenv==1.2.1
requests==2.32.5
rich==14.3.1
shellingham==1.5.4
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
